Изменено отображение текста азимута
• Теперь текст будет от 19.5 до 60, а потом от 0 до 40.5.
• Линия сдвинута часть дальше, чтобы не выпирала из блока.

+++ Assets/Scripts/Block/Control/Amperage1.cs
@@ -21,7 +21,7 @@ namespace Assets.Scripts.Block.Control
             _reflectorHandle.RemoveListener(Rotate);
         }
 
-        private void Rotate(int value)
+        private void Rotate(float value)
         {
             _center.localEulerAngles = new Vector3(-28 + value * 34 / 100, 0, 180);
         }
